[FIX] : 플레이어와 적 상호작용 구현 TakeDamage, Attack 상호작용
