+++ DarkShield/Assets/02.Scripts/Enemy/EleteEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31bbc0d0447478e4f988e7a19c133e80
+guid: b739870059d385841aaccca6830f92d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
