--- DarkShield/Assets/02.Scripts/Enemy/RangeEnemy/MelleBoss.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MelleBoss : NormalBoss
-{
-
-    public override void Attack()
-    {
-        float distanceToPlayer = Vector3.Distance(transform.position, target.position);
-        if (distanceToPlayer <= attackRange && Time.time >= lastAttackTime + attackCooldown)
-        {
-            MeleeAttack();
-        }
-    }
-
-    private void MeleeAttack()
-    {
-        Debug.Log("근거리 공격!");
-        // 근거리 공격 로직 (예: 데미지 적용)
-        lastAttackTime = Time.time;
-    }
-}
