--- DarkShield/Assets/02.Scripts/Enemy/RangeEnemy/MelleBoss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2eacd8ea948fdff42a19463301a8bc15
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
