+++ DarkShield/ProjectSettings/TagManager.asset
@@ -13,7 +13,7 @@ TagManager:
   - Water
   - UI
   - Portal
-  - 
+  - Enemy
   - 
   - 
   -
